app/requirements.txt
pymorphy2
beautifulsoup4
deep-translator
feedparser
psycopg[binary]
django-environ>=0.7,<1.0
django
gunicorn==20.0.4
lxml